Depth and traversal limits for aggregate size calculations
