+++ common/types/BUILD.bazel
@@ -29,6 +29,7 @@ go_library(
         "overflow.go",
         "provider.go",
         "regex.go",
+        "size_calc.go",
         "string.go",
         "struct.go",
         "timestamp.go",
@@ -77,6 +78,7 @@ go_test(
         "optional_test.go",
         "provider_test.go",
         "regex_test.go",
+        "size_calc_test.go",
         "string_test.go",
         "timestamp_test.go",
         "types_test.go",
